+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/IrContributionData.kt
@@ -3,13 +3,20 @@
 package dev.zacsweers.metro.compiler.ir
 
 import dev.zacsweers.metro.compiler.Symbols
+import dev.zacsweers.metro.compiler.flatMapToSet
 import dev.zacsweers.metro.compiler.mapNotNullToSet
 import dev.zacsweers.metro.compiler.mapToSet
+import org.jetbrains.kotlin.descriptors.Visibilities
 import org.jetbrains.kotlin.ir.declarations.IrClass
+import org.jetbrains.kotlin.ir.declarations.IrDeclarationWithVisibility
+import org.jetbrains.kotlin.ir.declarations.IrFile
+import org.jetbrains.kotlin.ir.declarations.moduleDescriptor
 import org.jetbrains.kotlin.ir.types.IrType
 import org.jetbrains.kotlin.ir.types.classOrFail
 import org.jetbrains.kotlin.ir.types.classOrNull
 import org.jetbrains.kotlin.ir.util.defaultType
+import org.jetbrains.kotlin.ir.util.file
+import org.jetbrains.kotlin.ir.util.getPackageFragment
 import org.jetbrains.kotlin.ir.util.nestedClasses
 import org.jetbrains.kotlin.name.ClassId
 
@@ -46,18 +53,31 @@ internal class IrContributionData(private val metroContext: IrMetroContext) {
     addAll(findExternalBindingContainerContributions(scope))
   }
 
+  private fun findVisibleContributionClassesForScopeInHints(scopeClassId: ClassId): Set<IrClass> {
+    val functionsInPackage =
+      metroContext.referenceFunctions(Symbols.CallableIds.scopeHint(scopeClassId))
+    val contributingClasses =
+      functionsInPackage
+        .filter {
+          if (it.owner.visibility == Visibilities.Internal) {
+            it.owner.isVisibleAsInternal(it.owner.file)
+          } else {
+            true
+          }
+        }
+        .mapToSet { contribution ->
+          // This is the single value param
+          contribution.owner.regularParameters.single().type.classOrFail.owner
+        }
+    return contributingClasses
+  }
+
   // TODO this may do multiple lookups of the same origin class if it contributes to multiple scopes
   //  something we could possibly optimize in the future.
   private fun findExternalContributions(scopeClassId: ClassId): Set<IrType> {
     return externalContributions.getOrPut(scopeClassId) {
-      val functionsInPackage =
-        metroContext.referenceFunctions(Symbols.CallableIds.scopeHint(scopeClassId))
-      val contributingClasses =
-        functionsInPackage.map { contribution ->
-          // This is the single value param
-          contribution.owner.regularParameters.single().type.classOrFail.owner
-        }
-      getScopedContributions(contributingClasses, scopeClassId)
+      val contributingClasses = findVisibleContributionClassesForScopeInHints(scopeClassId)
+      getScopedContributions(contributingClasses, scopeClassId, bindingContainersOnly = false)
     }
   }
 
@@ -65,24 +85,20 @@ internal class IrContributionData(private val metroContext: IrMetroContext) {
   //  something we could possibly optimize in the future.
   private fun findExternalBindingContainerContributions(scopeClassId: ClassId): Set<IrClass> {
     return externalBindingContainerContributions.getOrPut(scopeClassId) {
-      val functionsInPackage =
-        metroContext.referenceFunctions(Symbols.CallableIds.scopeHint(scopeClassId))
-      val contributingClasses =
-        functionsInPackage.map { contribution ->
-          // This is the single value param
-          contribution.owner.regularParameters.single().type.classOrFail.owner
+      val contributingClasses = findVisibleContributionClassesForScopeInHints(scopeClassId)
+      getScopedContributions(contributingClasses, scopeClassId, bindingContainersOnly = true)
+        .mapNotNullToSet {
+          it.classOrNull?.owner?.takeIf {
+            it.isAnnotatedWithAny(metroContext.symbols.classIds.bindingContainerAnnotations)
+          }
         }
-      getScopedContributions(contributingClasses, scopeClassId).mapNotNullToSet {
-        it.classOrNull?.owner?.takeIf {
-          it.isAnnotatedWithAny(metroContext.symbols.classIds.bindingContainerAnnotations)
-        }
-      }
     }
   }
 
   private fun getScopedContributions(
-    contributingClasses: List<IrClass>,
+    contributingClasses: Collection<IrClass>,
     scopeClassId: ClassId,
+    bindingContainersOnly: Boolean,
   ): Set<IrType> {
     val filteredContributions = contributingClasses.toMutableList()
 
@@ -99,18 +115,34 @@ internal class IrContributionData(private val metroContext: IrMetroContext) {
       }
 
     return filteredContributions
-      .flatMap { it.nestedClasses }
-      .mapNotNullToSet { nestedClass ->
-        val metroContribution =
-          nestedClass.findAnnotations(Symbols.ClassIds.metroContribution).singleOrNull()
-            ?: return@mapNotNullToSet null
-        val contributionScope =
-          metroContribution.scopeOrNull()
-            ?: error("No scope found for @MetroContribution annotation")
-        if (contributionScope == scopeClassId) {
-          nestedClass.defaultType
+      .let { contributions ->
+        if (bindingContainersOnly) {
+          contributions.filter {
+            it.isAnnotatedWithAny(metroContext.symbols.classIds.bindingContainerAnnotations)
+          }
         } else {
-          null
+          contributions.filterNot {
+            it.isAnnotatedWithAny(metroContext.symbols.classIds.bindingContainerAnnotations)
+          }
+        }
+      }
+      .flatMapToSet {
+        if (it.isAnnotatedWithAny(metroContext.symbols.classIds.bindingContainerAnnotations)) {
+          setOf(it.defaultType)
+        } else {
+          it.nestedClasses.mapNotNullToSet { nestedClass ->
+            val metroContribution =
+              nestedClass.findAnnotations(Symbols.ClassIds.metroContribution).singleOrNull()
+                ?: return@mapNotNullToSet null
+            val contributionScope =
+              metroContribution.scopeOrNull()
+                ?: error("No scope found for @MetroContribution annotation")
+            if (contributionScope == scopeClassId) {
+              nestedClass.defaultType
+            } else {
+              null
+            }
+          }
         }
       }
   }
@@ -138,5 +170,14 @@ internal class IrContributionData(private val metroContext: IrMetroContext) {
 
       return unfilteredScopedInjectClasses
     }
+  }
+
+  // Copied from CheckerUtils.kt
+  private fun IrDeclarationWithVisibility.isVisibleAsInternal(file: IrFile): Boolean {
+    val referencedDeclarationPackageFragment = getPackageFragment()
+    val module = file.module
+    return module.descriptor.shouldSeeInternalsOf(
+      referencedDeclarationPackageFragment.moduleDescriptor
+    )
   }
 }