+++ gradle-plugin/src/main/kotlin/dev/zacsweers/metro/gradle/MetroPluginExtension.kt
@@ -68,17 +68,21 @@ constructor(layout: ProjectLayout, objects: ObjectFactory, providers: ProviderFa
     objects.property(Boolean::class.javaObjectType).convention(false)
 
   /**
-   * Enable/disable hint property generation for scoped inject classes. Enabled by default.
+   * Enable/disable hint property generation for scoped inject classes. Disabled by default.
    *
    * @see <a href="https://zacsweers.github.io/metro/dependency-graphs/#graph-extensions">Graph
    *   Extensions docs for more details</a>
    */
   public val enableScopedInjectClassHints: Property<Boolean> =
-    objects.property(Boolean::class.javaObjectType).convention(true)
+    objects.property(Boolean::class.javaObjectType).convention(false)
 
   /** Enable/disable shrinking of unused bindings. Enabled by default. */
   public val shrinkUnusedBindings: Property<Boolean> =
     objects.property(Boolean::class.javaObjectType).convention(true)
+
+  /** Enable/disable chunking of field initializers. Disabled by default. */
+  public val chunkFieldInits: Property<Boolean> =
+    objects.property(Boolean::class.javaObjectType).convention(false)
 
   /** Enable/disable automatic transformation of providers to be private. Enabled by default. */
   public val transformProvidersToPrivate: Property<Boolean> =
@@ -101,6 +105,13 @@ constructor(layout: ProjectLayout, objects: ObjectFactory, providers: ProviderFa
       .convention(
         providers.gradleProperty("metro.version.check").map { it.toBoolean() }.orElse(true)
       )
+
+  /**
+   * Enable/disable suggestion to lift @Inject to class when there is only one constructor. Enabled
+   * by default.
+   */
+  public val warnOnInjectAnnotationPlacement: Property<Boolean> =
+    objects.property(Boolean::class.javaObjectType).convention(true)
 
   /**
    * If set, the Metro compiler will dump report diagnostics about resolved dependency graphs to the